+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - A way to append items to navbar component.
 - Strict type declaration everywhere.
+- Text input blade component.
 
 ### Changed
 - Refactor tests